requirements.txt
flask
openpyxl
numpy
pyodbc
gunicorn
SpeechRecognition
wave
deepgram-sdk
Translator
PyAudio